added Damagecomponent and added it to hero and monsters
Now there are damage components because there just weren't enough components already

+++ game/src/ecs/components/stats/XPModifier.java
@@ -1,4 +1,4 @@
 package ecs.components.stats;
 
-public record XPModifier(int xpMultiplier) {
+public record XPModifier(float xpMultiplier) {
 }
